internal/service/seckill_test.go
package service

import (
	"context"
	"testing"
	"time"
	
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/mock"
	"seckill/internal/model"
)

type MockProductRepo struct {
	mock.Mock
}

func (m *MockProductRepo) GetByID(ctx context.Context, id uint) (*model.Product, error) {
	args := m.Called(ctx, id)
	if args.Get(0) == nil {
		return nil, args.Error(1)
	}
	return args.Get(0).(*model.Product), args.Error(1)
}

// 其他mock方法实现...

func TestPlaceOrder(t *testing.T) {
	tests := []struct {
		name        string
		userID      uint
		activityID  uint
		mockSetup   func(*MockProductRepo)
		expectError bool
	}{
		{
			name:       "success",
			userID:     1,
			activityID: 1,
			mockSetup: func(repo *MockProductRepo) {
				repo.On("GetByID", mock.Anything, uint(1)).Return(&model.Product{
					ID:    1,
					Stock: 10,
				}, nil)
			},
			expectError: false,
		},
		// 添加更多测试用例...
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			mockRepo := new(MockProductRepo)
			tt.mockSetup(mockRepo)

			service := NewSeckillService(mockRepo, nil, nil, nil)
			err := service.PlaceOrder(context.Background(), tt.userID, tt.activityID)

			if tt.expectError {
				assert.Error(t, err)
			} else {
				assert.NoError(t, err)
			}
			mockRepo.AssertExpectations(t)
		})
	}
} 